windows_requirements.txt
anyio==3.4.0
asgiref==3.4.1
beautifulsoup4==4.9.3
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
fastapi==0.70.1
folium==0.12.1.post1
fonttools==4.28.5
googlesearch-python==1.0.1
h11==0.12.0
idna==2.10
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.6
PySimpleGUI==4.55.1
python-barcode==0.13.1
python-dateutil==2.8.2
requests==2.25.1
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
sounddevice==0.4.4
soupsieve==2.3.1
starlette==0.16.0
tk==0.1.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
